initial logic for key rotation
