+++ internal/keyrotation/server.go
@@ -19,13 +19,21 @@ import (
 	"context"
 	"fmt"
 	"net/http"
+	"sync"
 
-	"go.opencensus.io/trace"
-
-	"github.com/google/exposure-notifications-server/internal/logging"
 	revisiondb "github.com/google/exposure-notifications-server/internal/revision/database"
 	"github.com/google/exposure-notifications-server/internal/serverenv"
 )
+
+// Server hosts end points to manage key rotation
+type Server struct {
+	config     *Config
+	env        *serverenv.ServerEnv
+	revisionDB *revisiondb.RevisionDB
+
+	// mutex lock to be held over read/write operations when rotating keys.
+	mu sync.RWMutex
+}
 
 // NewServer creates a Server that manages deletion of
 // old export files that are no longer needed by clients for download.
@@ -53,13 +61,6 @@ func NewServer(config *Config, env *serverenv.ServerEnv) (*Server, error) {
 	}, nil
 }
 
-// Server hosts end points to manage key rotation
-type Server struct {
-	config     *Config
-	env        *serverenv.ServerEnv
-	revisionDB *revisiondb.RevisionDB
-}
-
 // Routes defines and returns the routes for this server.
 func (s *Server) Routes(ctx context.Context) *http.ServeMux {
 	mux := http.NewServeMux()
@@ -67,26 +68,4 @@ func (s *Server) Routes(ctx context.Context) *http.ServeMux {
 	mux.HandleFunc("/rotate-keys", s.handleRotateKeys(ctx))
 
 	return mux
-}
-
-func (s *Server) handleRotateKeys(ctx context.Context) http.HandlerFunc {
-	logger := logging.FromContext(ctx).Named("keyrotation.HandleRotate")
-
-	return func(w http.ResponseWriter, r *http.Request) {
-		_, span := trace.StartSpan(r.Context(), "(*keyrotation.handler).ServeHTTP")
-		defer span.End()
-
-		// TODO(whaught):
-		// 1. Retrieve keys
-		// 2. Early exit if the newest is new (configurable)
-		// 3. Take a lock on the DB
-		// 4. Otherwise generate a key and store
-		// 5. delete oldest keys, but always keep any within 15d or still primary
-		// 6. Metric on count created and deleted
-		// 7. logger.log that too
-		// 8. Unlock
-
-		logger.Info("key rotation complete.")
-		w.WriteHeader(http.StatusOK)
-	}
 }
